Fix bssid parsing
When using a comma separated bssid arg the program would try to open it
as a file and fail. This has been fixed by trying to open the file. If
the file exists it proceeds if not it tries to split on commas.

Also fixed bssids not returning in slice format